requirements.txt
numpy==1.24.4
pillow==10.2.0
pydantic==2.6.3
requests==2.31.0
